chore: renamed screens
