--- src/lib/utils/error.ts
@@ -1,4 +1,0 @@
-export function throwError(e: unknown): never {
-	console.error({ e })
-	throw e
-}
